ILEX-64 connect together add new term and term dialogs
