+++ src/components/SingleTermView/OverView/PredicateGroupInput.jsx
@@ -8,7 +8,6 @@ import SingleSearch from "../SingleSearch";
 import { debounce } from 'lodash';
 import predicatesData from "../../../static/predicates.json"
 import {getMatchTerms} from "../../../api/endpoints";
-import {ToggleButton, ToggleButtonGroup} from "@mui/lab";
 import * as mockApi from '../../../api/endpoints/swaggerMockMissingEndpoints';
 
 const {gray700, gray300, gray800, gray600} = vars
