+++ src/components/TermEditor/ImportFileTab.jsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { Box, Stack, Typography, FormControl, Divider } from "@mui/material";
 import CustomizedInput from "../common/CustomizedInput";
 import ImportFile from "./ImportFile";
