+++ src/components/TermEditor/TermDialog.jsx
@@ -1,26 +1,16 @@
 import * as React from "react";
-import { useState, useEffect, useMemo, useCallback } from "react";
-import { Box, Divider, MobileStepper, Stack, Button, Typography, Chip } from "@mui/material";
+import { useState } from "react";
+import { Box, Divider, MobileStepper, Stack, Button } from "@mui/material";
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
 import { vars } from "../../theme/variables";
 import CustomizedDialog from "../common/CustomizedDialog";
 import CustomButton from "../common/CustomButton";
-import TermForm from "./TermForm";
-import TermSidebar from "./TermSidebar";
-import AddPredicatesStep from "./AddPredicatesStep";
-import TermStatusStep from "./TermStatusStep";
-import * as mockApi from "../../api/endpoints/swaggerMockMissingEndpoints";
-import * as mockApiInterlex from "../../api/endpoints/interLexURIStructureAPI";
-import { termParser } from "../../parsers/termParser";
-import { useQuery } from "../../helpers";
-import { debounce } from 'lodash';
+import TermDialogContent from "./TermDialogContent";
+import AddNewTermDialogContent from "./AddNewTermDialogContent";
 
-const useMockApi = () => mockApi;
-const useMockApiInterlex = () => mockApiInterlex;
+const { gray100, gray200, gray400, brand700 } = vars;
 
-const { gray100, gray200, gray400, brand700, success600, success700 } = vars;
-
-const HeaderRightSideContent = ({ activeStep, onContinue, onClose }) => (
+const HeaderRightSideContent = ({ activeStep, onContinue, onClose, isContinueButtonDisabled }) => (
     <Box display='flex' alignItems='center'>
         {activeStep !== 2 ? (
             <>
@@ -42,6 +32,7 @@ const HeaderRightSideContent = ({ activeStep, onContinue, onClose }) => (
                     <CustomButton onClick={onClose}>Cancel</CustomButton>
                     <Button
                         onClick={onContinue}
+                        disabled={isContinueButtonDisabled}
                         variant="contained"
                         endIcon={<ArrowForwardIcon />}
                         sx={{
@@ -59,112 +50,21 @@ const HeaderRightSideContent = ({ activeStep, onContinue, onClose }) => (
     </Box>
 );
 
-const TermDialog = ({ open, handleClose }) => {
-    const { getMatchTerms } = useMockApi();
-    const { getEndpointsIlx } = useMockApiInterlex();
-    const query = useQuery();
-    const searchTerm = query.get('searchTerm');
-    const [loading, setLoading] = useState(true);
-    const [termResults, setTermResults] = useState([]);
+const TermDialog = ({ open, handleClose, searchTerm }) => {
     const [activeStep, setActiveStep] = useState(0);
-    const [openSidebar, setOpenSidebar] = useState(true);
-    const [responseStatus, setResponseStatus] = useState('success')
-    const [predicates, setPredicates] = useState([{ subject: '', predicate: '', object: { type: 'Object', value: '', isLink: false } }])
-    const [data, setData] = useState(null);
-    const [formState, setFormState] = useState({
-        label: searchTerm || '',
-        age: '',
-        synonyms: [],
-        superclass: '',
-        existingIds: [],
-        urls: '',
-        description: '',
-        comment: ''
-    });
+    const [areMatchesChecked, setAreMatchesChecked] = useState(false);
 
-    const fetchTerms = useCallback(
-        debounce((searchTerm) => {
-            setLoading(true);
-            if (searchTerm) {
-                getMatchTerms("base", searchTerm).then(data => {
-                    const parsedData = termParser(data);
-                    setData(parsedData?.results[0]);
-                    setLoading(false);
-                });
-            } else {
-                getMatchTerms("base", "i", { filter: "", value: "" }).then(data => {
-                    const parsedData = termParser(data, "");
-                    setTermResults(parsedData.results);
-                    setLoading(false);
-                });
-            }
-        }, 300),
-        [getEndpointsIlx, getMatchTerms]
-    );
 
-    const handleContinueClick = () => {
-        setActiveStep(activeStep + 1);
-        if (activeStep === 2) {
-            console.log("POST: here connect to post request")
-            //here we change status as well according to api response
-        }
-    }
-    const handleSidebarToggle = () => setOpenSidebar(!openSidebar);
-    const handleCancelBtnClick = () => { handleClose(); setActiveStep(0); setAreMatchesChecked(false); };
-    const handleAddNewTerm = () => { setActiveStep(0); setAreMatchesChecked(false); };
-    const handleFormInputChange = (e) => {
-        const { name, value } = e.target;
-        if (name === "label") {
-            setTermValue(e.target.value);
-        }
-        setFormState((prevState) => ({
-            ...prevState,
-            [name]: value
-        }));
-    };
-
-    const handleAutocompleteChange = (name) => (event, value) => {
-        setFormState((prevState) => ({
-            ...prevState,
-            [name]: value
-        }));
-    };
-
-    useEffect(() => {
-        fetchTerms(searchTerm);
-        return () => {
-            fetchTerms.cancel();
-        };
-    }, [searchTerm, fetchTerms]);
-
-    useEffect(() => {
-        if (data) {
-            setFormState((prevState) => ({
-                ...prevState,
-                synonyms: data.synonym || [],
-                existingIds: data.existingID || [],
-                description: data.description || ''
-            }));
-        }
-    }, [data]);
-
-    const memoData = useMemo(() => data, [data]);
-
-    useEffect(() => {
-        if (memoData?.predicates) {
-            setPredicates(memoData.predicates);
-        }
-    }, [memoData]);
-
-    const predicatesOptions = predicates.map(row => ({
-        label: row.title,
-        value: row.title
-    }));
-
+    const handleContinueClick = () => { setActiveStep(activeStep + 1) }
+    const handleMatchesChange = (e) => setAreMatchesChecked(e.target.checked);
+    const handleCancelBtnClick = () => { handleClose(); setActiveStep(0); setAreMatchesChecked(false) };
+    const handleReset = () => { setAreMatchesChecked(false); setActiveStep(0) }
+    const isSearchTermAvailable = searchTerm ? true : false;
+    const isContinueButtonDisabled = !isSearchTermAvailable && !areMatchesChecked;
 
     return (
         <CustomizedDialog
-            title={`Suggest changes to “${searchTerm}”`}
+            title={searchTerm ? `Suggest changes to “${searchTerm}”` : 'Add a new term'}
             open={open}
             handleClose={handleClose}
             HeaderRightSideContent={
@@ -172,36 +72,25 @@ const TermDialog = ({ open, handleClose }) => {
                     activeStep={activeStep}
                     onContinue={handleContinueClick}
                     onClose={handleCancelBtnClick}
+                    isContinueButtonDisabled={isContinueButtonDisabled}
                 />
             }
             sx={{ '& .MuiDialogContent-root': { padding: 0, overflowY: "hidden" } }}
         >
-            {activeStep === 0 && (
-                <Box display="flex" height={1}>
-                    <Box sx={{ px: '3.25rem', pt: '1.75rem', pb: '2.5rem', flex: 1, overflowY: 'auto' }}>
-                        <Stack direction="row" gap={1.5} alignItems="center">
-                            <Typography variant="h5" sx={{ fontWeight: 500 }}>{searchTerm}</Typography>
-                            <Chip
-                                label="Active"
-                                sx={{
-                                    border: `1.5px solid ${success600}`,
-                                    background: 'transparent',
-                                    color: success700
-                                }}
-                            />
-                        </Stack>
-                        <TermForm
-                            formState={formState}
-                            data={memoData}
-                            onInputChange={handleFormInputChange}
-                            onAutocompleteChange={handleAutocompleteChange}
-                        />
-                    </Box>
-                    <TermSidebar open={openSidebar} loading={loading} onToggle={handleSidebarToggle} results={termResults} data={memoData} />
-                </Box>
+            {searchTerm ? (
+                <TermDialogContent
+                    activeStep={activeStep}
+                    searchTerm={searchTerm}
+                    onReset={handleReset}
+                />
+            ) : (
+                <AddNewTermDialogContent
+                    activeStep={activeStep}
+                    areMatchesChecked={areMatchesChecked}
+                    onMatchesChange={handleMatchesChange}
+                    onReset={handleReset}
+                />
             )}
-            {activeStep === 1 && <AddPredicatesStep searchTerm={searchTerm} predicatesOptions={predicatesOptions} />}
-            {activeStep === 2 && <TermStatusStep responseStatus={responseStatus} termValue={searchTerm} onAddNewTerm={handleAddNewTerm} />}
         </CustomizedDialog>
     );
 };
